Publish Test PyPI releases on every commit to main (#21)
This modifies the `publish.yml` workflow to run with every commit to
`main` branch. These commits will then lead to producing a Test PyPI
release.

That way, we will get to test out our publishing CI constantly, so that
when we do want to push out a release, we don't need to be afraid that
it might not work since the CI was modified and this is the first time
it's running.

Additionally, it will give our users the option to use these TestPyPI
releases if they wish a more up-to-date version of the library, without
having to result to directly using git and building the package locally.
This also gives users the benefit of having checkable attestations,
proving that the build files were produced by our CI on the TestPyPI.

Note: This modifies `uv-dynamic-versioning` to avoid including the
commit sha as the local version component (e.g. instead of
`0.6.0.post10.dev0+81baa11`, only produce `0.6.0.post10.dev0`). This is
because PyPI [doesn't
support](https://packaging.python.org/en/latest/specifications/version-specifiers/#version-specifiers)
the use of local version components.